+++ NEWS.md
@@ -1,3 +1,12 @@
+# rUM 1.0
+
+* Added quarto options in project creation template/wizard
+* Added quarto support for make_project
+* Added rStudio project files
+* Added check to make sure analysis does not exist - Thanks to Francisco Cardozo
+* Updated .gitignore
+* Added citation info to README
+
 # rUM 0.3.2
 
 * Fixed package version for table1 to be `packageVersion("table1")` instead of `packageVersion("tidyverse")`